fixed nuiInit/nuiUninit problems (the kernel was not properly initialized) better VST tutorial GUI fixed nuiClampedValueAttributeEditor
git-svn-id: svn://libnui.net/nui/trunk/nui3@4980 9db30882-7939-4ce0-a36a-45f8a15c01b5

+++ include/nuiClampedValueAttributeEditor.h
@@ -147,13 +147,13 @@ protected:
     
     if (mIndexed)
     {
-      mAttribute.Set(mIndex, mpKnob->GetRange().GetValue());
+      mAttribute.Set(mIndex, (float)mpKnob->GetRange().GetValue());
       mAttribute.Format(mIndex, str);
       setValue = mAttribute.Get(mIndex);
     }
     else
     {
-      mAttribute.Set(mpKnob->GetRange().GetValue());
+      mAttribute.Set((float)mpKnob->GetRange().GetValue());
       mAttribute.Format(str);
       setValue = mAttribute.Get();
     }
